+++ postgres/CHANGES.md
@@ -1,6 +1,7 @@
 # unreleased 0.3.0
 ## Remove 
 - remove `ExecuteMut` trait. It's role is replaced by `impl Execute<&mut C>`
+- remove `error::AuthenticationError` type. It's error condition is covered by `error::ConfigError`
 
 ## Change
 - change `pool::Pool`'s dead connection detection lifecycle.